requirements.txt
streamlit==1.37.1
google-generativeai==0.7.2
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.2
plotly==5.23.0
orjson==3.10.7
